#1935 Converted ViolationElement back to record, moved relaxing annotations to the annotation package.
