--- platform-annotations/src/main/java/ua/com/fielden/platform/annotations/ClassLevelAnnotation.java
@@ -1,7 +1,0 @@
-package ua.com.fielden.platform.annotations;
-
-import java.lang.annotation.*;
-
-@Target(value = { ElementType.TYPE })
-public @interface ClassLevelAnnotation {
-}
